\name{brbPackageTry-package}
\alias{brbPackageTry-package}
\alias{brbPackageTry}
\docType{package}
\title{
Berkeley R Beginners Package Try
}
\description{
This is an example package created during a meeting of the Berkeley R Beginner's
group at the University of California, Berkeley.
}
\details{
\tabular{ll}{
Package: \tab brbPackageTry\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-03-17\cr
License: \tab CC0\cr
}
}
\author{
Allen Hammock
}
\references{
http://r-pkgs.had.co.nz/
}
\keyword{ package }
\seealso{
}
\examples{
hello()
hello("Charlie")
add(2, 2)
}
