% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say someone's name.}
\usage{
hello(who = "Anonymous Person")
}
\arguments{
\item{who}{A person's name.}
}
\value{
A text string containing the \code{who}'s name.
}
\description{
This is an example function that prints 'Hello, <who>!' or
'Hello, Anonymous Person!' if no argument is given.
}
\examples{
hello("Charlie")
hello()
}

